put building service in a common builder
