+++ packages/push-signal-service/src/services/queque.service.ts
@@ -2,12 +2,7 @@ import { Logger, SQS } from "signalhub-commons";
 import { config } from "../config/config.js";
 import { signalNotSendedToQueque } from "../model/domain/errors.js";
 
-const sqsClient: SQS.SQSClient = SQS.instantiateClient({
-  region: config.region,
-  endpoint: config.queueEndpoint,
-});
-
-export function quequeServiceBuilder() {
+export function quequeServiceBuilder(sqsClient: SQS.SQSClient) {
   return {
     async send(message: string, logger: Logger): Promise<void> {
       try {
